pyproject.toml
[project]
name = "chponskiy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-compressor>=4.5.1",
    "django-debug-toolbar>=5.1.0",
    "django-libsass>=0.9",
    "django-widget-tweaks>=1.5.0",
    "gunicorn>=23.0.0",
    "pydantic>=2.11.2",
    "pydantic-settings>=2.8.1",
]

[dependency-groups]
dev = [
    "pylint>=3.3.6",
    "pylint-django>=2.6.1",
]
